requirements.txt
aiogram>=3.0
pydantic>=2.0
python-dotenv
httpx
groq
